refactor: 요구사항과 다른 API에서 받아왔던 DTO를 맞게 수정함
