+++ WeatherForecast/WeatherForecast/Controller/ViewController.swift
@@ -22,6 +22,4 @@ class ViewController: UIViewController {
             }
         }
     }
-    
 }
-
