+++ WeatherForecast/WeatherForecast/Model/DTO/Forecast.swift
@@ -1,5 +1,5 @@
 //
-//  DailyForecast.swift
+//  Forecast.swift
 //  WeatherForecast
 //
 //  Created by DONGWOOK SEO on 2023/03/13.
@@ -7,10 +7,11 @@
 
 import Foundation
 
-// MARK: - DailyForecast
-struct DailyForecast: Decodable {
+// MARK: - Forecast
+struct Forecast: Decodable {
     let cod: String
-    let message, cnt: Int
+    let message: Int
+    let cnt: Int
     let list: [List]
     let city: City
 }
@@ -21,7 +22,10 @@ struct City: Decodable {
     let name: String
     let coord: Coord
     let country: String
-    let population, timezone, sunrise, sunset: Int
+    let population: Int
+    let timezone: Int
+    let sunrise: Int
+    let sunset: Int
 }
 
 // MARK: - List
@@ -31,12 +35,22 @@ struct List: Decodable {
     let weather: [Weather]
     let clouds: Clouds
     let wind: Wind
-    let visibility, pop: Int
+    let visibility: Int
+    let pop: Double
     let dtTxt: String
+    let rain: Rain?
 
     enum CodingKeys: String, CodingKey {
-        case dt, main, weather, clouds, wind, visibility, pop
+        case dt, main, weather, clouds, wind, visibility, pop, rain
         case dtTxt = "dt_txt"
     }
 }
 
+// MARK: - Rain
+struct Rain: Decodable {
+    let the3H: Double
+
+    enum CodingKeys: String, CodingKey {
+        case the3H = "3h"
+    }
+}
